Tools to support DE analysis

Tools to support DE analysis

level_matcher(raw_levels = NULL, ordered_levels = NULL)

UpdateLevelsbyList(metadata = NULL, named_levels = NULL)

valid_factors(metadata = NULL)

Arguments

raw_levels

string of factor levels in data to be sorted.

ordered_levels

string of factor levels in the desired order

metadata

A data.frame containing factors

named_levels

a named list containing the levels to set as reference via fct_relevel

Value

data.frame with factors

Details

level_matcher() returns overlapping strings between raw_levels and ordered_levels, with the ordering of the later. No partial string matching. This is useful when applying a generic factor level order to different metadata that do not share all the same levels.

The names from named_levels map to the columns in metadata to update. List slots contain strings passed to the ... argument of fct_relevel

This is all done with imap_dfc

drops single level factors with metadata select_if(~n_distinct(na.omit(.x))>1)