Support functions for GSEA

`OntologyMethods()` and `OntologyFunctions()` work together to supply geneset library functions for the `Shiny_Hotgenes()` app.

The `msigdbr_wrapper()` function is a wrapper that is used supply the default Ontology_Function argument for `OntologyMethods()`, although it can be used as a standalone function for supplying genesets.

fgsea_(
  Ranks = NULL,
  pathways = NULL,
  minSize = 1,
  maxSize = Inf,
  nproc = 1,
  ...
)

fgsea_Results(
  fgseaRes = NULL,
  Topn = Inf,
  padj_cut = 0.1,
  contrasts = NULL,
  mode = "Details"
)

GSEA_Plots(
  fgseaRes = NULL,
  Topn = 10,
  padj_cut = 0.1,
  contrasts = NULL,
  width = 50
)

leadingGenes(fgseaRes = NULL, contrast = NULL, genesetName = NULL)

plotEnrichment_(fgseaRes = NULL, contrast = NULL, genesetName = NULL)

OntologyMethods(
  Ontology_Function = list(msigdbr = msigdbr_wrapper),
  InputChoices = list(msigdbr = msigdbr_wrapper_choices()$set),
  gene_col_choices = list(msigdbr = c("gene_symbol", "entrez_gene", "ensembl_gene")),
  species_choices = list(msigdbr = c("human", "mouse", "rat", "dog")),
  versions = list(msigdbr = packageVersion("msigdbr"))
)

OntologyFunctions(
  Methods = NULL,
  db = NULL,
  species = NULL,
  set = NULL,
  gene_col = NULL,
  ...
)

msigdbr_wrapper_choices()

msigdbr_wrapper(
  species = "human",
  set = NULL,
  gene_col = NULL,
  clean_names = TRUE
)

Arguments

Ranks

A named list containing gene stats in descending order.

pathways

List of gene sets to check.

minSize

Minimal size of a gene set to test. All pathways below the threshold are excluded.

maxSize

Maximal size of a gene set to test. All pathways above the threshold are excluded.

nproc

If not equal to zero sets BPPARAM to use nproc workers (default = 0).

...

additional args passed to supplied methods

fgseaRes

A named list containing the output from fgsea_.

Topn

number indicating the top number of pathways to show. Results grouped by sign(NES) and then Topn max absolute NES selected.

padj_cut

numeric value for padj limit (0.1 is Default).

contrasts

vector for DE data.frames in Output_DE slot to return. See ?contrasts_. If NULL (default), all contrasts used.

mode

vector indicating the details to return. Options include, Details (Complete pathway details per contrast), leadingEdge (vector for of the column with the same name Length (Number pathways per contrast),

width

Positive integer giving target line width (in number of characters). A width less than or equal to 1 will put each word on its own line.

contrast

String indicating the contrast to select.

genesetName

String indicating the pathway to return.

Ontology_Function

a named list containing functions for returning a named list of geneset for GSVA/GSEA analysis.

InputChoices

a named list containing a vector with string choices corresponding to ontology choices for supplied ontology functions.

gene_col_choices

a named list containing a vector with string choices corresponding to gene id choices for supplied ontology functions.

species_choices

a named list containing a vector with string choices corresponding to species choices for supplied ontology functions.

versions

a named list containing Ontology database version number as a reference.

Methods

output of OntologyMethods function

db

string indicating the ontology database to use.

species

maps to the species argument of msigdbr.

set

maps to the gs_subcat argument of msigdbr, when not an empty string. If gs_subcat is an empty string, uses gs_cat. Multiple okay.

gene_col

string specifying id to use. Choices include "gene_symbol" (default),

clean_names

Logical, if TRUE (default) geneset names are passed to make_clean_names

Details

`fgsea_()` detects for ranks that are all positive or all negative and then switches to "pos" or "neg", depending on the outcome. Otherwise scoreType is "std"

`msigdbr_wrapper()` returns a named list containing genesets provided by msigdbr. List object names are supplied by the "gs_name" column and are populated by the column specify by the gene_col argument.

Examples

if(interactive()){
  # Getting example data ----------------------------------------------------
  
  # load package
  library(Hotgenes)
  
  dds_Hotgenes_dir <- system.file("extdata",
                                  paste0("dds_Hotgenes", ".RDS"),
                                  package = "Hotgenes",
                                  mustWork = TRUE
  )
  # from DESeq2
  htgs <- readRDS(dds_Hotgenes_dir)
  
  # load signatures using msigdbr wrapper
  # This returns a named list of pathways
  
  H_paths <- msigdbr_wrapper(
    species = "human",
    set = c("CP:REACTOME", "CP:KEGG"),
    gene_col = "gene_symbol"
  )
  
  H_paths %>% names() %>% head()
  # custom geneset functions can be supplied like this
  # This can be used in the Shiny_Hotgenes() function
  Custom_geneset_function_list <- OntologyMethods(
    Ontology_Function = list("msigdbr" = msigdbr_wrapper),
    InputChoices = list("msigdbr" = c("CP:REACTOME", "CP:KEGG")),
    gene_col_choices = list("msigdbr" = c(
      "gene_symbol", "entrez_gene", "ensembl_gene"
    )),
    species_choices = list("msigdbr" = c("human", "mouse", "rat", "dog")),
    versions = list("msigdbr" = packageVersion("msigdbr"))
  )
  
  
  msigdbr_pthyways <- OntologyFunctions(
    Methods = Custom_geneset_function_list,
    db = "msigdbr",
    species = "human",
    set = c("CP:REACTOME", "CP:KEGG"),
    gene_col = "gene_symbol"
  )
  
  testthat::expect_equal(H_paths, msigdbr_pthyways)
  
  # These genes sets are mapped to gene symbols
  # Verify Feature col contains gene symbols, too
  # In this example the "Feature" column contains gene symbols
  
  htgs %>% Mapper_()
  
  # Get ranks
  InputRanks <- htgs %>%
    DE(
      Report = "Ranks",
      contrasts = "Hrs_2_vs_0",
      Rank_name = "Feature", # see above
      padj_cut = 1
    )
  
  
  # fgsea wrapper --------
  Out_GSEA <- fgsea_(
    Ranks = InputRanks,
    pathways = H_paths,
    nproc = 8,
    minSize = 5,
    maxSize = Inf
  )
  
  # Get details for all
  Out_GSEA %>%
    fgsea_Results(
      padj_cut = 0.2,
      mode = "D"
    )
  
  # Or for one
  Out_GSEA %>%
    fgsea_Results(
      contrasts = "Hrs_2_vs_0",
      padj_cut = 0.2,
      mode = "leadingEdge"
    )
  
  # Generate a summary plot
  Out_GSEA %>%
    GSEA_Plots(
      contrasts = "Hrs_2_vs_0",
      padj_cut = 0.2,
      width = 30,
      Topn = 2
    )
  
  # plotEnrichment_
  plotEnrichment_(
    Out_GSEA, "Hrs_2_vs_0",
    "kegg_nod_like_receptor_signaling_pathway"
  )
  
  
  # leadingGenes
  leadingGenes(
    Out_GSEA, "Hrs_2_vs_0",
    "kegg_nod_like_receptor_signaling_pathway"
  )
  
}