Hotgenes class

Hotgenes class

Returns features from auxiliary_assays slot

Returns colnames for coldata slot

auxiliary_assays_features(Hotgenes = NULL)

auxiliary_assays_default(Hotgenes = NULL, coldata_ids = NULL)

auxiliary_assays_(Hotgenes = NULL, features = NULL, mode = "all")

auxiliary_assays_(object) <- value

# S4 method for Hotgenes
auxiliary_assays_(object) <- value

O_(Hotgenes = NULL)

coldata_(Hotgenes = NULL, coldata_ids = coldata_names(Hotgenes), mode = "all")

coldata_names(Hotgenes = NULL)

coldata_(object) <- value

# S4 method for Hotgenes
coldata_(object) <- value

designMatrix_(Hotgenes = NULL)

contrastMatrix_(Hotgenes = NULL)

Normalized_Data_(Hotgenes = NULL, slot = NULL)

SampleIDs_(Hotgenes = NULL)

ExpressionSlots_(Hotgenes = NULL)

ExpressionData_(Hotgenes = NULL, slot = ExpressionSlots_(Hotgenes)[1])

Output_DE_(
  Hotgenes = NULL,
  contrasts = NULL,
  mapFeatures = TRUE,
  hotList = NULL,
  padj_cut = 0.1,
  .log2FoldChange = 0,
  annotateSig = TRUE,
  as_list = FALSE
)

contrasts_(Hotgenes = NULL)

Features_(Hotgenes = NULL)

Mapper_(Hotgenes = NULL)

Mapper_(object) <- value

# S4 method for Hotgenes
Mapper_(object) <- value

Background_(Hotgenes = NULL, Col = "Feature")

Arguments

Hotgenes

Hotgenes object.

coldata_ids

Variable(s) stored in coldata slot can be selected by column name, using a vector strings. Default is coldata_names(). If set to "", an empty data.frame is returned.

features

vector of feature to return. If NULL (default) all data will be returned. No need to specifiy "SampleIDs".

mode

string indicating the type of coldata to call. Choices include: "all", "quanti", "quali". If "all", the default, all columns will be returned. If "quanti", only numeric coldata will be returned. If "quali", only qualitative coldata will be returned.

object

Hotgenes object.

value

a data.frame containing a "Feature" for alias mapping.

slot

Single name of expression data.frame in Normalized_Expression slot to be returned. Default is first string returned by ExpressionSlots_()

contrasts

vector for DE data.frames in Output_DE slot to return. See ?contrasts_. If NULL (default), all contrasts used.

mapFeatures

If FALSE, no changes will be made. If TRUE (default), features will be mapped to information provided in the mapper slot. See details below and Mapper_.

hotList

vector of Features to select. This will overide all cut offs.

padj_cut

numeric value for padj limit (0.1 is Default).

.log2FoldChange

absolute log2FoldChange for filtering. Default is 0.

annotateSig

logical, if TRUE (default), a column is added to show significance for any hotList feature (based on cut offs).

as_list

logical, default is FALSE. If true, results returned as a named list other object with be a dataframe grouped by "contrast".

Col

String matching the column id from Mapper slot to pull. Default is "Feature".

Value

auxiliary_assays_default() returns a data.frame with the minimum requirements for a Hotgenes object. The coldata_ids parameter can be used to define metadata columns to include in the data.frame, to help align data.

auxiliary_assays_ returns a data.frame containing auxiliary assays to complement expression data

O_ Returns the original R object imported into Hotgenes. For HotgenesDESeq2 this a DESeqDataSet class object. For HotgenesMonocle3 this is a cell_data_set class object. For Hotgeneslimma this is an EList class object (Expression data), with the limmafit (MArrayLM class object) add in in the "limmafit" slot.

coldata_ Returns Coldata (Experimental Variables).

designMatrix_ Returns the model matrix used for the original DE analysis (if available).

contrastMatrix_ Returns the contrast matrix used for the original DE analysis (if available).

Named list from Normalized_Expression slot if "slot" is NULL. If slot argument is provided, a data.frame is returned.

SampleIDs_ Returns the colnames of the expression data in the first ExpressionSlot.

ExpressionSlots_ Returns the names for all the available expression data. This depends on the original R package used for DE analysis. For example, DESeq2 data can be normalized by "rld" or "VST".

returns a data.frame

Output_DE_ Returns complete unfiltered Output_DE slot object.

contrasts_ Returns exported contrasts.

Features_ Returns the rownames of the expression data in the first ExpressionSlot. It will also check for empty rownames and will return an error if any are detected. This will be used as a last check before returning a hotgenes object during generation.

Mapper_ Returns Mapper containing aliases mapped to Features. Only aliases in expression data are returned

Background_ Returns a vector of all Features in the expression data.

Details

Except for when mode = "all", single level columns will be dropped.

Slots

Output_DE

Named list of data.frames Each data.frame contains atleast: "Feature", "baseMean","log2FoldChange", "stat", "pvalue", "padj". Additional values dependent on original DE analysis. Accessible via Output_DE_.

Normalized_Expression

named list of data.frames available expression data. Names of available slots accessible via ExpressionSlots_. Expression data can be accessed via Normalized_Data_.

coldata

The coldata, as a data.frame with Sample Metadata. Accessible via coldata_.

designMatrix

The model matrix used for DE analysis. Accessible via designMatrix_.

contrastMatrix

The contrast matrix used for DE analysis. Accessible via contrastMatrix_.

Original_Object

This is the original object imported into Hotgenes. Accessible via O_.

auxiliary_assays

A data.frame with auxiliary assays

Mapper

Slot for a data.frame that can be used to map aliases to original assay names. Must be a data.frame containing a "Feature" column, which will be used for mapping to results. This object can be updated or viewed using Mapper_

Examples

# load package
library(Hotgenes)

fit_Hotgenes_dir <- system.file("extdata",
  paste0("fit_Hotgenes", ".RDS"),
  package = "Hotgenes",
  mustWork = TRUE
)

# from limma
fit_Hotgenes <- readRDS(fit_Hotgenes_dir) %>% 
  update_object()

# call the object to get summary
fit_Hotgenes
#> class: Hotgenes 
#> Original class/package:  EList/limma
#> 
#> Differential expression (default thresholds): 
#> |contrast       | total|
#> |:--------------|-----:|
#> |Hrs_2_vs_0     |    45|
#> |Hrs_6_vs_0     |    71|
#> |sh_EWS_vs_Ctrl |    51|
#> |shEWS.Hrs2     |     8|
#> |shEWS.Hrs6     |     1|
#> 
#> Available feature mapping:  Feature, ensembl_id 
#> ExpressionSlots:  logCPM, counts 
#> Total auxiliary assays:  0 
#> Total samples:  12 

# auxiliary_assays_ -------------------------------------------------------
fit_Hotgenes %>% 
auxiliary_assays_()
#> data frame with 0 columns and 12 rows

# ExpressionSlots_ ------------------------------------
fit_Hotgenes %>% ExpressionSlots_()
#> [1] "logCPM" "counts"

# Normalized_Data_ --------------------------------------------------------
# should return object class data.frame when slot specified
fit_Hotgenes %>%
  Normalized_Data_(slot = "logCPM") %>% head()
#>        shCON_0hrs_1 shCON_0hrs_2 shCON_2hrs_1 shCON_2hrs_2 shCON_6hrs_1
#> AGER       5.699205     6.356261     6.030818     5.461677     6.150731
#> ALOX12     3.377277     3.372749     3.472823     3.422148     3.298289
#> ALOX15     5.699205     5.795961     5.512351     5.206420     5.172758
#> ALOX5      4.294815     4.305635     2.624826     5.007111     4.215826
#> AREG       5.822062     5.528028     6.185541     5.806812     6.194453
#> ARG1       5.822062     5.795961     5.160879     5.461677     5.413766
#>        shCON_6hrs_2 shEWS_0hrs_1 shEWS_0hrs_2 shEWS_2hrs_1 shEWS_2hrs_2
#> AGER       5.601095     6.519313     6.607402     6.174031     6.441586
#> ALOX12     4.635860     4.324297     4.379596     3.777141     3.666292
#> ALOX15     5.735396     5.748795     5.341122     5.296515     5.638277
#> ALOX5      3.413468     3.776809     4.219131     3.329682     3.666292
#> AREG       6.352067     5.691079     5.686257     5.484960     6.155677
#> ARG1       4.887399     5.630958     5.489221     5.598171     5.457705
#>        shEWS_6hrs_1 shEWS_6hrs_2
#> AGER       5.903504     5.960532
#> ALOX12     4.229732     3.391689
#> ALOX15     5.263679     5.317689
#> ALOX5      4.592302     3.877116
#> AREG       6.803199     6.868503
#> ARG1       4.791611     5.157224

# O_ --------------------------------------------------
# get original object
fit_Hotgenes %>%
  O_() %>%
  class()
#> [1] "EList"
#> attr(,"package")
#> [1] "limma"

# designMatrix_ ---------------------------------------
fit_Hotgenes %>%
  designMatrix_() %>%
  head()
#>              Intercept sh_EWS_vs_Ctrl Hrs_2_vs_0 Hrs_6_vs_0 shEWS.Hrs2
#> shCON_0hrs_1         1              0          0          0          0
#> shCON_0hrs_2         1              0          0          0          0
#> shCON_2hrs_1         1              0          1          0          0
#> shCON_2hrs_2         1              0          1          0          0
#> shCON_6hrs_1         1              0          0          1          0
#> shCON_6hrs_2         1              0          0          1          0
#>              shEWS.Hrs6
#> shCON_0hrs_1          0
#> shCON_0hrs_2          0
#> shCON_2hrs_1          0
#> shCON_2hrs_2          0
#> shCON_6hrs_1          0
#> shCON_6hrs_2          0

# coldata_ --------------------------------------------
fit_Hotgenes %>% 
  coldata_()
#>                sh Bio_Rep Hrs Time
#> shCON_0hrs_1 Ctrl       1   0    0
#> shCON_0hrs_2 Ctrl       2   0    0
#> shCON_2hrs_1 Ctrl       1   2    2
#> shCON_2hrs_2 Ctrl       2   2    2
#> shCON_6hrs_1 Ctrl       1   6    6
#> shCON_6hrs_2 Ctrl       2   6    6
#> shEWS_0hrs_1  EWS       1   0    0
#> shEWS_0hrs_2  EWS       2   0    0
#> shEWS_2hrs_1  EWS       1   2    2
#> shEWS_2hrs_2  EWS       2   2    2
#> shEWS_6hrs_1  EWS       1   6    6
#> shEWS_6hrs_2  EWS       2   6    6

# SampleIDs_ ------------------------------------------
fit_Hotgenes %>% SampleIDs_()
#>  [1] "shCON_0hrs_1" "shCON_0hrs_2" "shCON_2hrs_1" "shCON_2hrs_2" "shCON_6hrs_1"
#>  [6] "shCON_6hrs_2" "shEWS_0hrs_1" "shEWS_0hrs_2" "shEWS_2hrs_1" "shEWS_2hrs_2"
#> [11] "shEWS_6hrs_1" "shEWS_6hrs_2"

# contrasts_ ------------------------------------------
fit_Hotgenes %>% contrasts_()
#> [1] "Hrs_2_vs_0"     "Hrs_6_vs_0"     "sh_EWS_vs_Ctrl" "shEWS.Hrs2"    
#> [5] "shEWS.Hrs6"    

# Mapper_ ---------------------------------------------
fit_Hotgenes %>%
  Mapper_() %>%
  head()
#> # A tibble: 6 × 2
#>   Feature ensembl_id     
#>   <chr>   <chr>          
#> 1 AGER    ENSG00000204305
#> 2 AGER    ENSG00000230514
#> 3 AGER    ENSG00000229058
#> 4 AGER    ENSG00000231268
#> 5 AGER    ENSG00000237405
#> 6 AGER    ENSG00000234729

# Features_ -------------------------------------------
fit_Hotgenes %>% Features_()
#>   [1] "AGER"         "ALOX12"       "ALOX15"       "ALOX5"        "AREG"        
#>   [6] "ARG1"         "ATF2"         "BCL2L1"       "BCL6"         "BIRC2"       
#>  [11] "BORCS8-MEF2B" "C1QA"         "C1QB"         "C1R"          "C1S"         
#>  [16] "C2"           "C3"           "C3AR1"        "C4A"          "C5"          
#>  [21] "C6"           "C7"           "C8A"          "C8B"          "C9"          
#>  [26] "CCL11"        "CCL13"        "CCL16"        "CCL17"        "CCL19"       
#>  [31] "CCL2"         "CCL20"        "CCL21"        "CCL22"        "CCL23"       
#>  [36] "CCL24"        "CCL3"         "CCL4"         "CCL5"         "CCL7"        
#>  [41] "CCL8"         "CCR1"         "CCR2"         "CCR3"         "CCR4"        
#>  [46] "CCR7"         "CD163"        "CD4"          "CD40"         "CD40LG"      
#>  [51] "CD55"         "CD86"         "CDC42"        "CEBPB"        "CFB"         
#>  [56] "CFD"          "CFL1"         "CLTC"         "CREB1"        "CRP"         
#>  [61] "CSF1"         "CSF2"         "CSF3"         "CXCL1"        "CXCL10"      
#>  [66] "CXCL2"        "CXCL3"        "CXCL5"        "CXCL6"        "CXCL8"       
#>  [71] "CXCL9"        "CXCR1"        "CXCR2"        "CXCR4"        "CYSLTR1"     
#>  [76] "CYSLTR2"      "DAXX"         "DDIT3"        "DEFA1"        "ELK1"        
#>  [81] "FASLG"        "FLT1"         "FOS"          "FXYD2"        "GAPDH"       
#>  [86] "GNAQ"         "GNAS"         "GNB1"         "GNGT1"        "GRB2"        
#>  [91] "GUSB"         "HDAC4"        "HIF1A"        "HLA-DRA"      "HLA-DRB1"    
#>  [96] "HMGB1"        "HMGB2"        "HMGN1"        "HPRT1"        "HRAS"        
#> [101] "HSH2D"        "HSPB1"        "HSPB2"        "IFI44"        "IFIT1"       
#> [106] "IFIT2"        "IFIT3"        "IFNA1"        "IFNB1"        "IFNG"        
#> [111] "IL10"         "IL10RB"       "IL11"         "IL12A"        "IL12B"       
#> [116] "IL13"         "IL15"         "IL17A"        "IL18"         "IL18RAP"     
#> [121] "IL1A"         "IL1B"         "IL1R1"        "IL1RAP"       "IL1RN"       
#> [126] "IL2"          "IL21"         "IL22"         "IL22RA2"      "IL23A"       
#> [131] "IL23R"        "IL3"          "IL4"          "IL5"          "IL6"         
#> [136] "IL6R"         "IL7"          "IL9"          "IRF1"         "IRF3"        
#> [141] "IRF5"         "IRF7"         "ITGB2"        "JUN"          "KEAP1"       
#> [146] "KNG1"         "LIMK1"        "LTA"          "LTB"          "LTB4R"       
#> [151] "LTB4R2"       "LY96"         "MAFF"         "MAFG"         "MAFK"        
#> [156] "MAP2K1"       "MAP2K4"       "MAP2K6"       "MAP3K1"       "MAP3K5"      
#> [161] "MAP3K7"       "MAP3K9"       "MAPK1"        "MAPK14"       "MAPK3"       
#> [166] "MAPK8"        "MAPKAPK2"     "MAPKAPK5"     "MASP1"        "MASP2"       
#> [171] "MAX"          "MBL2"         "MEF2A"        "MEF2C"        "MEF2D"       
#> [176] "MKNK1"        "MMP3"         "MMP9"         "MRC1"         "MX1"         
#> [181] "MX2"          "MYC"          "MYD88"        "MYL2"         "NFATC3"      
#> [186] "NFE2L2"       "NFKB1"        "NLRP3"        "NOD1"         "NOD2"        
#> [191] "NOS2"         "NOX1"         "NR3C1"        "OAS2"         "OASL"        
#> [196] "OXER1"        "PDGFA"        "PGK1"         "PIK3C2G"      "PLA2G4A"     
#> [201] "PLCB1"        "PPP1R12B"     "PRKCA"        "PRKCB"        "PTGDR2"      
#> [206] "PTGER1"       "PTGER2"       "PTGER3"       "PTGER4"       "PTGFR"       
#> [211] "PTGIR"        "PTGS1"        "PTGS2"        "PTK2"         "RAC1"        
#> [216] "RAF1"         "RAPGEF2"      "RELA"         "RELB"         "RHOA"        
#> [221] "RIPK1"        "RIPK2"        "ROCK2"        "RPS6KA5"      "SHC1"        
#> [226] "SMAD7"        "STAT1"        "STAT2"        "STAT3"        "TBXA2R"      
#> [231] "TCF4"         "TGFB1"        "TGFB2"        "TGFB3"        "TGFBR1"      
#> [236] "TLR1"         "TLR2"         "TLR3"         "TLR4"         "TLR5"        
#> [241] "TLR6"         "TLR7"         "TLR8"         "TLR9"         "TNF"         
#> [246] "TNFAIP3"      "TNFSF14"      "TOLLIP"       "TRADD"        "TRAF2"       
#> [251] "TREM2"        "TSLP"         "TUBB"         "TWIST2"       "TYROBP"      

# Background_ -----------------------------------------
# this returns Features mapped to Mapper aliases
fit_Hotgenes %>% 
  Background_(Col = "ensembl_id")
#>   [1] "ENSG00000204305" "ENSG00000230514" "ENSG00000229058" "ENSG00000231268"
#>   [5] "ENSG00000237405" "ENSG00000234729" "ENSG00000206320" "ENSG00000108839"
#>   [9] "ENSG00000161905" "ENSG00000012779" "ENSG00000275565" "ENSG00000109321"
#>  [13] "ENSG00000118520" "ENSG00000115966" "ENSG00000171552" "ENSG00000113916"
#>  [17] "ENSG00000110330" "ENSG00000213999" "ENSG00000173372" "ENSG00000173369"
#>  [21] "ENSG00000159403" "ENSG00000288512" "ENSG00000182326" "ENSG00000166278"
#>  [25] "ENSG00000206372" "ENSG00000231543" "ENSG00000204364" "ENSG00000226560"
#>  [29] "ENSG00000235696" "ENSG00000235017" "ENSG00000125730" "ENSG00000171860"
#>  [33] "ENSG00000244731" "ENSG00000227746" "ENSG00000206340" "ENSG00000244207"
#>  [37] "ENSG00000106804" "ENSG00000039537" "ENSG00000112936" "ENSG00000157131"
#>  [41] "ENSG00000021852" "ENSG00000113600" "ENSG00000172156" "ENSG00000181374"
#>  [45] "ENSG00000275152" "ENSG00000275095" "ENSG00000102970" "ENSG00000172724"
#>  [49] "ENSG00000108691" "ENSG00000115009" "ENSG00000137077" "ENSG00000102962"
#>  [53] "ENSG00000274736" "ENSG00000276114" "ENSG00000106178" "ENSG00000277632"
#>  [57] "ENSG00000278567" "ENSG00000274221" "ENSG00000275302" "ENSG00000277943"
#>  [61] "ENSG00000275824" "ENSG00000271503" "ENSG00000274233" "ENSG00000108688"
#>  [65] "ENSG00000108700" "ENSG00000163823" "ENSG00000121807" "ENSG00000183625"
#>  [69] "ENSG00000183813" "ENSG00000126353" "ENSG00000177575" "ENSG00000010610"
#>  [73] "ENSG00000101017" "ENSG00000102245" "ENSG00000196352" "ENSG00000114013"
#>  [77] "ENSG00000070831" "ENSG00000172216" "ENSG00000243649" "ENSG00000242335"
#>  [81] "ENSG00000241534" "ENSG00000204359" "ENSG00000243570" "ENSG00000241253"
#>  [85] "ENSG00000239754" "ENSG00000197766" "ENSG00000274619" "ENSG00000172757"
#>  [89] "ENSG00000141367" "ENSG00000118260" "ENSG00000132693" "ENSG00000184371"
#>  [93] "ENSG00000164400" "ENSG00000108342" "ENSG00000163739" "ENSG00000169245"
#>  [97] "ENSG00000081041" "ENSG00000163734" "ENSG00000163735" "ENSG00000124875"
#> [101] "ENSG00000169429" "ENSG00000138755" "ENSG00000163464" "ENSG00000180871"
#> [105] "ENSG00000121966" "ENSG00000173198" "ENSG00000152207" "ENSG00000204209"
#> [109] "ENSG00000231617" "ENSG00000229396" "ENSG00000227046" "ENSG00000206206"
#> [113] "ENSG00000206279" "ENSG00000175197" "ENSG00000206047" "ENSG00000284983"
#> [117] "ENSG00000126767" "ENSG00000117560" "ENSG00000102755" "ENSG00000170345"
#> [121] "ENSG00000137731" "ENSG00000111640" "ENSG00000156052" "ENSG00000087460"
#> [125] "ENSG00000078369" "ENSG00000127928" "ENSG00000177885" "ENSG00000169919"
#> [129] "ENSG00000068024" "ENSG00000100644" "ENSG00000204287" "ENSG00000277263"
#> [133] "ENSG00000234794" "ENSG00000226260" "ENSG00000230726" "ENSG00000227993"
#> [137] "ENSG00000228987" "ENSG00000206308" "ENSG00000196126" "ENSG00000229074"
#> [141] "ENSG00000228080" "ENSG00000206306" "ENSG00000236884" "ENSG00000206240"
#> [145] "ENSG00000189403" "ENSG00000164104" "ENSG00000205581" "ENSG00000165704"
#> [149] "ENSG00000174775" "ENSG00000276536" "ENSG00000196684" "ENSG00000106211"
#> [153] "ENSG00000170276" "ENSG00000137965" "ENSG00000185745" "ENSG00000119922"
#> [157] "ENSG00000119917" "ENSG00000197919" "ENSG00000171855" "ENSG00000111537"
#> [161] "ENSG00000136634" "ENSG00000243646" "ENSG00000095752" "ENSG00000168811"
#> [165] "ENSG00000113302" "ENSG00000169194" "ENSG00000164136" "ENSG00000112115"
#> [169] "ENSG00000150782" "ENSG00000115607" "ENSG00000115008" "ENSG00000125538"
#> [173] "ENSG00000115594" "ENSG00000196083" "ENSG00000136689" "ENSG00000109471"
#> [177] "ENSG00000138684" "ENSG00000127318" "ENSG00000164485" "ENSG00000110944"
#> [181] "ENSG00000162594" "ENSG00000164399" "ENSG00000113520" "ENSG00000113525"
#> [185] "ENSG00000136244" "ENSG00000160712" "ENSG00000104432" "ENSG00000145839"
#> [189] "ENSG00000125347" "ENSG00000126456" "ENSG00000128604" "ENSG00000185507"
#> [193] "ENSG00000276561" "ENSG00000160255" "ENSG00000177606" "ENSG00000079999"
#> [197] "ENSG00000113889" "ENSG00000106683" "ENSG00000226979" "ENSG00000230279"
#> [201] "ENSG00000238130" "ENSG00000231408" "ENSG00000173503" "ENSG00000223919"
#> [205] "ENSG00000226275" "ENSG00000227507" "ENSG00000223448" "ENSG00000231314"
#> [209] "ENSG00000204487" "ENSG00000238114" "ENSG00000236925" "ENSG00000236237"
#> [213] "ENSG00000206437" "ENSG00000213903" "ENSG00000285456" "ENSG00000213906"
#> [217] "ENSG00000285203" "ENSG00000154589" "ENSG00000185022" "ENSG00000197063"
#> [221] "ENSG00000198517" "ENSG00000169032" "ENSG00000065559" "ENSG00000108984"
#> [225] "ENSG00000095015" "ENSG00000197442" "ENSG00000135341" "ENSG00000006432"
#> [229] "ENSG00000100030" "ENSG00000112062" "ENSG00000102882" "ENSG00000107643"
#> [233] "ENSG00000162889" "ENSG00000089022" "ENSG00000127241" "ENSG00000009724"
#> [237] "ENSG00000125952" "ENSG00000165471" "ENSG00000068305" "ENSG00000081189"
#> [241] "ENSG00000116604" "ENSG00000079277" "ENSG00000149968" "ENSG00000100985"
#> [245] "ENSG00000260314" "ENSG00000157601" "ENSG00000183486" "ENSG00000136997"
#> [249] "ENSG00000172936" "ENSG00000111245" "ENSG00000072736" "ENSG00000116044"
#> [253] "ENSG00000109320" "ENSG00000162711" "ENSG00000106100" "ENSG00000167207"
#> [257] "ENSG00000007171" "ENSG00000007952" "ENSG00000113580" "ENSG00000111335"
#> [261] "ENSG00000135114" "ENSG00000162881" "ENSG00000197461" "ENSG00000102144"
#> [265] "ENSG00000139144" "ENSG00000116711" "ENSG00000182621" "ENSG00000077157"
#> [269] "ENSG00000154229" "ENSG00000166501" "ENSG00000183134" "ENSG00000160951"
#> [273] "ENSG00000125384" "ENSG00000050628" "ENSG00000171522" "ENSG00000122420"
#> [277] "ENSG00000160013" "ENSG00000095303" "ENSG00000073756" "ENSG00000169398"
#> [281] "ENSG00000136238" "ENSG00000132155" "ENSG00000109756" "ENSG00000173039"
#> [285] "ENSG00000104856" "ENSG00000067560" "ENSG00000137275" "ENSG00000104312"
#> [289] "ENSG00000134318" "ENSG00000100784" "ENSG00000160691" "ENSG00000101665"
#> [293] "ENSG00000115415" "ENSG00000170581" "ENSG00000168610" "ENSG00000006638"
#> [297] "ENSG00000196628" "ENSG00000105329" "ENSG00000092969" "ENSG00000119699"
#> [301] "ENSG00000106799" "ENSG00000174125" "ENSG00000137462" "ENSG00000164342"
#> [305] "ENSG00000136869" "ENSG00000187554" "ENSG00000174130" "ENSG00000196664"
#> [309] "ENSG00000101916" "ENSG00000239732" "ENSG00000232810" "ENSG00000228978"
#> [313] "ENSG00000204490" "ENSG00000223952" "ENSG00000228849" "ENSG00000228321"
#> [317] "ENSG00000230108" "ENSG00000206439" "ENSG00000118503" "ENSG00000125735"
#> [321] "ENSG00000078902" "ENSG00000102871" "ENSG00000127191" "ENSG00000095970"
#> [325] "ENSG00000145777" "ENSG00000196230" "ENSG00000224156" "ENSG00000235067"
#> [329] "ENSG00000229684" "ENSG00000232575" "ENSG00000232421" "ENSG00000183311"
#> [333] "ENSG00000227739" "ENSG00000233608" "ENSG00000288335" "ENSG00000011600"

# updating object ---------------------------------------------------------

# Adding column of strings
coldata_(fit_Hotgenes)$New <- "value"
coldata_(fit_Hotgenes)
#>                sh Bio_Rep Hrs Time   New
#> shCON_0hrs_1 Ctrl       1   0    0 value
#> shCON_0hrs_2 Ctrl       2   0    0 value
#> shCON_2hrs_1 Ctrl       1   2    2 value
#> shCON_2hrs_2 Ctrl       2   2    2 value
#> shCON_6hrs_1 Ctrl       1   6    6 value
#> shCON_6hrs_2 Ctrl       2   6    6 value
#> shEWS_0hrs_1  EWS       1   0    0 value
#> shEWS_0hrs_2  EWS       2   0    0 value
#> shEWS_2hrs_1  EWS       1   2    2 value
#> shEWS_2hrs_2  EWS       2   2    2 value
#> shEWS_6hrs_1  EWS       1   6    6 value
#> shEWS_6hrs_2  EWS       2   6    6 value

# update mapper slot

# in case you wanted to include aliases for your genes
# requires a "Feature" column that contains gene names in expression matrix
dbCon <- org.Hs.eg.db::org.Hs.eg_dbconn()

sqlQuery <- "SELECT * FROM ENSEMBL, gene_info WHERE ENSEMBL._id == gene_info._id;"

ensembl_Symbol <- DBI::dbGetQuery(dbCon, sqlQuery) %>%
  dplyr::select(dplyr::any_of(c("Feature" = "symbol", 
                                "gene_name", 
                                "ensembl_id"))) %>%
  tibble::as_tibble()

Test_Hotgenes <- fit_Hotgenes

Mapper_(Test_Hotgenes) <- ensembl_Symbol
Mapper_(Test_Hotgenes)
#> # A tibble: 339 × 3
#>    Feature gene_name                                            ensembl_id     
#>    <chr>   <chr>                                                <chr>          
#>  1 AGER    advanced glycosylation end-product specific receptor ENSG00000204305
#>  2 AGER    advanced glycosylation end-product specific receptor ENSG00000229058
#>  3 AGER    advanced glycosylation end-product specific receptor ENSG00000234729
#>  4 AGER    advanced glycosylation end-product specific receptor ENSG00000231268
#>  5 AGER    advanced glycosylation end-product specific receptor ENSG00000230514
#>  6 AGER    advanced glycosylation end-product specific receptor ENSG00000206320
#>  7 AGER    advanced glycosylation end-product specific receptor ENSG00000237405
#>  8 ALOX12  arachidonate 12-lipoxygenase, 12S type               ENSG00000108839
#>  9 ALOX15  arachidonate 15-lipoxygenase                         ENSG00000161905
#> 10 ALOX5   arachidonate 5-lipoxygenase                          ENSG00000012779
#> # ℹ 329 more rows

# auxiliary_assays_ --------------------------------------------------------
fit_Hotgenes_2 <- fit_Hotgenes

coldata_(fit_Hotgenes_2)$Samples <- as.character(SampleIDs_(fit_Hotgenes_2))

df_assays_slot <- auxiliary_assays_default(fit_Hotgenes_2) 
df_assays_slot
#>       SampleIDs
#> 1  shCON_0hrs_1
#> 2  shCON_0hrs_2
#> 3  shCON_2hrs_1
#> 4  shCON_2hrs_2
#> 5  shCON_6hrs_1
#> 6  shCON_6hrs_2
#> 7  shEWS_0hrs_1
#> 8  shEWS_0hrs_2
#> 9  shEWS_2hrs_1
#> 10 shEWS_2hrs_2
#> 11 shEWS_6hrs_1
#> 12 shEWS_6hrs_2

# making example data
set.seed(12)

max_len <- length(SampleIDs_(fit_Hotgenes_2))

AssayData  <- df_assays_slot %>% 
  dplyr::mutate(assay1 = rnorm(max_len),
                assay2 = rnorm(max_len))

AssayData
#>       SampleIDs     assay1      assay2
#> 1  shCON_0hrs_1 -1.4805676 -0.77956651
#> 2  shCON_0hrs_2  1.5771695  0.01195176
#> 3  shCON_2hrs_1 -0.9567445 -0.15241624
#> 4  shCON_2hrs_2 -0.9200052 -0.70346425
#> 5  shCON_6hrs_1 -1.9976421  1.18887916
#> 6  shCON_6hrs_2 -0.2722960  0.34051227
#> 7  shEWS_0hrs_1 -0.3153487  0.50696817
#> 8  shEWS_0hrs_2 -0.6282552 -0.29330515
#> 9  shEWS_2hrs_1 -0.1064639  0.22364142
#> 10 shEWS_2hrs_2  0.4280148  2.00720146
#> 11 shEWS_6hrs_1 -0.7777196  1.01197912
#> 12 shEWS_6hrs_2 -1.2938823 -0.30245925

auxiliary_assays_(fit_Hotgenes_2) <- AssayData
auxiliary_assays_(fit_Hotgenes_2) 
#>                  assay1      assay2
#> shCON_0hrs_1 -1.4805676 -0.77956651
#> shCON_0hrs_2  1.5771695  0.01195176
#> shCON_2hrs_1 -0.9567445 -0.15241624
#> shCON_2hrs_2 -0.9200052 -0.70346425
#> shCON_6hrs_1 -1.9976421  1.18887916
#> shCON_6hrs_2 -0.2722960  0.34051227
#> shEWS_0hrs_1 -0.3153487  0.50696817
#> shEWS_0hrs_2 -0.6282552 -0.29330515
#> shEWS_2hrs_1 -0.1064639  0.22364142
#> shEWS_2hrs_2  0.4280148  2.00720146
#> shEWS_6hrs_1 -0.7777196  1.01197912
#> shEWS_6hrs_2 -1.2938823 -0.30245925