Creates a new Hotgenes object from user-provided objects.

Creates a new Hotgenes object from user-provided objects.

HotgenesUniversal(
  Output_DE = NULL,
  Normalized_Expression = NULL,
  auxiliary_assays = NULL,
  coldata = NULL,
  Original_Object = NULL,
  designMatrix = NULL,
  contrastMatrix = NULL,
  Mapper = NULL
)

Arguments

Output_DE

(Required). A named list of data.frames. Each data.frame contains atleast: "Feature", "baseMean","log2FoldChange", "stat", "pvalue", "padj".

Normalized_Expression

(Required). A named list of data.frames available expression data. Columns must be sample names and rows must be features.

auxiliary_assays

(Optional). A data.frame with a SampleIDs column matching the rownames of coldata object. This data.frame can be used to store auxiliary data, such as clinical scores, observations, and/or low-throughput assay data. See details below.

coldata

(Required). A data.frame with sample metadata. Rownames must match column names of expression data. Columns are design varaibles or sample metadata.

Original_Object

(Optional). This is the original DE object used for DE analysis, such as limma fit object.

designMatrix

(Required). The model matrix or comparison used for DE analysis.

contrastMatrix

(optional). The contrast matrix used for DE analysis, if different from designMatrix.

Mapper

(Required). A data.frame that can be used to map aliases to original assay names. Must have a "Feature" column, which will be used for mapping to results. This object can be updated or viewed using Mapper_.

Examples

if(interactive()){
  # load packages
  library(Hotgenes)
  
  
  # load example data -------------------------------------------------------
  
  dds_Hotgenes_dir <- system.file("extdata",
                                  paste0("dds_Hotgenes", ".RDS"),
                                  package = "Hotgenes",
                                  mustWork = TRUE
  )
  
  htgs <- readRDS(dds_Hotgenes_dir)
  
  # creating new Hotgenes object --------------------------------------------
  
  # Getting example named list of DE statistics
  NewDE <- Output_DE_(htgs, as_list = TRUE, padj_cut = 1)
  NewDE %>% purrr::map(~ head(.))
  
  # Getting example named list of normalized data
  NormlData <- Normalized_Data_(htgs)
  NormlData %>% purrr::map(~ head(.))
  
  # Getting example coldata
  ExpColdata <- coldata_(htgs)
  ExpColdata
  
  # Getting example original data object used for DE analysis
  # This example was generated from DESeq2
  OrigDEObj <- O_(htgs)
  OrigDEObj %>% class()
  
  # Getting example design matrix
  DE_design <- designMatrix_(htgs)
  DE_design
  
  # Getting example mapper
  MapperDF <- Mapper_(htgs)
  MapperDF %>% head()
  
  # Converting example objects to hotgenes
  Hotgenes_Object <- HotgenesUniversal(
    Output_DE = NewDE,
    Normalized_Expression = NormlData,
    coldata = ExpColdata,
    Original_Object = OrigDEObj,
    designMatrix = DE_design,
    Mapper = MapperDF
  )
  
  Hotgenes_Object
  
  
}