tabPanel_Hotgenes() provides UI for loaded Hotgenes object

tabPanel_Hotgenes() provides UI for loaded Hotgenes object

Modular Hotgenes UI

R shiny app for visualizing DE analysis

tabPanel_Hotgenes(
  id = "hotgenes_load",
  tabsetPanel_id = "tabset_Hotgenes",
  title = "Hotgenes",
  value = title
)

Shiny_Hotgenes_UI(id = NULL, tabsetPanel_id = "tabs", width = 2)

Shiny_Hotgenes_Server(
  id = NULL,
  input = NULL,
  output = NULL,
  session = NULL,
  Hotgenes = NULL,
  OntologyMethods = Hotgenes::OntologyMethods(),
  Mapper_choices = names(Mapper_(Hotgenes)),
  parallel.sz = 1L,
  max_col_levels = 12,
  selected_fillby = "",
  remove_modal_spinner = TRUE
)

Shiny_Hotgenes(
  Hotgenes = NULL,
  OntologyMethods = Hotgenes::OntologyMethods(),
  Mapper_choices = names(Mapper_(Hotgenes)),
  theme = NULL,
  parallel.sz = 1L,
  max_col_levels = 12
)

# S3 method for default
Shiny_Hotgenes(
  Hotgenes = NULL,
  OntologyMethods = Hotgenes::OntologyMethods(),
  Mapper_choices = names(Mapper_(Hotgenes)),
  theme = NULL,
  parallel.sz = 1L,
  max_col_levels = 12
)

# S3 method for list
Shiny_Hotgenes(
  Hotgenes = NULL,
  OntologyMethods = Hotgenes::OntologyMethods(),
  Mapper_choices = names(Mapper_(Hotgenes)),
  theme = NULL,
  parallel.sz = 1L,
  max_col_levels = 12
)

Arguments

id

An ID string that corresponds with the ID used to call the module's UI function.

tabsetPanel_id

string for tabsetPanel id prefix to use. Default is "tabs". This is what controls the visibility of sidepanels.

title

Display title for tab

value

The value that should be sent when tabsetPanel reports that this tab is selected. If omitted and tabsetPanel has an id, then the title will be used.

width

The width of the sidebar and main panel. By default, the sidebar takes up 1/3 of the width, and the main panel 2/3. The total width must be 12 or less.

input

shiny server input parameter

output

shiny server output parameter

session

shiny server session parameter

Hotgenes

R Hotgenes object

OntologyMethods

output of OntologyMethods function.

Mapper_choices

vector containing names of Mapper_ columns that should be used to map expression data to ontology gene identifiers. Default is names(Mapper_(Hotgenes)). A named vector may be used, as well.

parallel.sz

Number of threads of execution to use when doing the calculations in parallel.

max_col_levels

integer setting the maximum number of levels a variable, from the coldata, can have for plots.

selected_fillby

string for to be passed to BoxPlot as initial selection. Default is NULL.

remove_modal_spinner

logical, if TRUE remove_modal_spinner will run at the end of initial server run.

theme

One of the following:

NULL (the default), which implies a "stock" build of Bootstrap 3.

A bslib::bs_theme() object. This can be used to replace a stock build of Bootstrap 3 with a customized version of Bootstrap 3 or higher.

A character string pointing to an alternative Bootstrap stylesheet (normally a css file within the www directory, e.g. www/bootstrap.css).

Value

a shiny object for visualization.

a shiny UI and server for visualization.

Examples

library(Hotgenes)

# load example data
dds_Hotgenes_dir <- system.file("extdata",
  paste0("dds_Hotgenes", ".RDS"),
  package = "Hotgenes",
  mustWork = TRUE
)

# from DESeq2
HotgenesObj <- readRDS(dds_Hotgenes_dir) %>% Hotgenes::update_object()

fit_Hotgenes_dir <- system.file("extdata",
                                paste0("fit_Hotgenes", ".RDS"),
                                package = "Hotgenes",
                                mustWork = TRUE
)

# from limma
fit_Hotgenes <- readRDS(fit_Hotgenes_dir) 

# converting to list of Hotgenes
HotgenesList <- list(
  DESeq2_Hotgenes = HotgenesObj,
  limma_Hotgenes = fit_Hotgenes
) 

# app with single Hotgenes object
if(FALSE){
  Shiny_Hotgenes(fit_Hotgenes,
                 theme = "shinythemes/css/united.min.css")
}

# app with multiple Hotgenes objects
if(FALSE){
 
  Shiny_Hotgenes(HotgenesList)
  
}

# or gsva -----------------------------------------------------------------
if(FALSE){
# msigdbr_wrapper ---------------------------------------------------------
Hotgenes::msigdbr_wrapper_choices() 

choice_set <- c("CP:KEGG", "CP:REACTOME", "CP:WIKIPATHWAYS")
choice_id <- "gene_symbol"

gsList <- msigdbr_wrapper(
  species = "human",
  set = choice_set,
  gene_col = choice_id
)

# HotgeneSets -------------------------------------------------------------
HotgenesObj %>% Mapper_()

HotgenesObj %>% contrastMatrix_()

# ?HotgeneSets
HotgeneSets_out <- HotgeneSets(
  Hotgenes = HotgenesObj,
  gset.idx.list = gsList,
  kcdf = "Gaussian",
  method = "gsva",
  min.sz = 5,
  max.sz = Inf,
  parallel.sz = 16L
)

HotgeneSets_out

# converting to list of Hotgenes
Hotgenes_sets_List <- list(
  DESeq2_Hotgenes = HotgenesObj,
  HotgeneSets = HotgeneSets_out
) 

# app with multiple Hotgenes objects

  
  Shiny_Hotgenes(Hotgenes_sets_List,
                 theme = "shinythemes/css/united.min.css")
  
}