shiny plot download handler

shiny plot download handler

shiny plot download handler

interactive plot modules

interactive plot modules

UI_download_plot(id = NULL, label = "PDF")

Server_download_plot(
  id = NULL,
  reactive_plot = NULL,
  filename = "Figure",
  extension = ".pdf",
  units = "in",
  input,
  output,
  session,
  width = 10,
  height = 8,
  ...
)

UI_interactive_plot(id, width = "100%", height = "400px", ...)

Server_interactive_plot(
  id,
  input,
  output,
  session,
  refine_cols = NULL,
  reactive_plot = NULL,
  xvar = reactive(NULL),
  yvar = reactive(NULL),
  execOnResize = FALSE
)

Arguments

id

An ID string that corresponds with the ID used to call the module's UI function.

label

The label that should appear on the button.

reactive_plot

a reactive ggplot object.

filename

string for file name without extension. Default is"Figure"

extension

string for file extension. Default is".png"

units

One of the following units in which the width and height arguments are expressed: "in", "cm", "mm" or "px".

input

shiny server input parameter

output

shiny server output parameter

session

Session from which to make a child scope (the default should almost always be used).

width, height

Plot size in units expressed by the units argument. If not supplied, uses the size of the current graphics device.

...

additional arguments for plotOutput

refine_cols

string indicating which columns to show. If NULL (default), all cols will be shown.

xvar, yvar

A string giving the name of the variable on the x or y axis. These are only required for base graphics, and must be the name of a column in df.

execOnResize

If FALSE (the default), then when a plot is resized, Shiny will replay the plot drawing commands with grDevices::replayPlot() instead of re-executing expr. This can result in faster plot redrawing, but there may be rare cases where it is undesirable. If you encounter problems when resizing a plot, you can have Shiny re-execute the code on resize by setting this to TRUE.