Hotgenes volcano plot

Hotgenes volcano plot

VPlot(
  Hotgenes = NULL,
  contrasts = NULL,
  hotList = NULL,
  padj_cut = 0.1,
  .log2FoldChange = 0,
  Hide_labels = TRUE,
  max.overlaps = 50,
  repel_labels = TRUE,
  interactive = FALSE,
  col = c("#A6CEE3", "#1F78B4", "#B2DF8A", "#33A02C"),
  shinymode = FALSE,
  ...
)

Arguments

Hotgenes

Hotgenes object.

contrasts

vector for DE data.frames in Output_DE slot to return. See ?contrasts_. If NULL (default), all contrasts used.

hotList

vector of Features to select. This will overide all cut offs.

padj_cut

numeric value for padj limit (0.1 is Default).

.log2FoldChange

absolute log2FoldChange for filtering. Default is 0.

Hide_labels

Logical, if TRUE (Default), labels will not be shown.

max.overlaps

Exclude text labels when they overlap too many other things. For each text label, we count how many other text labels or other data points it overlaps, and exclude the text label if it has too many overlaps. Defaults to 10.

repel_labels

logical, if TRUE (default), labels will be generated using ggrepel::geom_text_repel. If FALSE, ggplot2::geom_text will be used.

interactive

logical, if TRUE, plotly object returned. If FALSE (default), ggplot object will be returned.

col

vector of colors to use for plot. Default is c("#A6CEE3", "#1F78B4", "#B2DF8A", "#33A02C") and maps with c("NS", "log2FC", "padj", "log2FC & padj").

shinymode

logical, if TRUE a named list of parameters will be exported to processing by internal plotting functions. If FALSE, a volcano plot is returned.

...

additional internal parameters.

Examples

if(interactive()){
  # load packages
  library(Hotgenes)
  
  dds_Hotgenes_dir <- system.file("extdata",
                                  paste0("dds_Hotgenes", ".RDS"),
                                  package = "Hotgenes",
                                  mustWork = TRUE
  )
  # from DESeq2
  htgs <- readRDS(dds_Hotgenes_dir)
  
  # VPlot -------------------------------------------------------------------
  
  # see all contrasts in object summary
  htgs %>% Mapper_()
  
  # select by name
 htgs %>%
    VPlot(
      contrasts = "sh_EWS_vs_Ctrl",
      .log2FoldChange = 0.5,
      hotList = "C1S",
      Hide_labels = FALSE,
      point_label_size = 5,
      base_size = 20,
      padj_cut = 0.1
    )

  # select by name
 
  htgs %>%
    VPlot(
      contrasts = "sh_EWS_vs_Ctrl",
      .log2FoldChange = 0.5,
      interactive = TRUE,
      hotList = "CCL2",
      Hide_labels = FALSE,
      pointSize = 2,
      max.overlaps = 50,
      legend_pointSize = 50,
      point_label_size = 5,
      padj_cut = 0.1
    ) 
  
}